web/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
pydantic>=2.5.0
python-dotenv>=1.0.0
duckdb>=1.1.0
